Support to flatten protobuf-struct deserialization
* Support to flatten protobuf-struct deserialization

This will convert protobuf-struct desrialization in to a clojure
persistent map. Protobuf structs are currently deserialized in to a
fields vector containing key value fields. This feature will add support
to flatten the vector containing key value fields in to a persistent map
recursively. To support backward compatibility, multi arity functions
are used that will take a boolean field to determine whether to flatten
a serialized protobuf struct.

Co-authored-by: Kishan Sharma <kishansharmacs@gmail.com>